The center of the zero looks bad in NTSC
NTSC artifacts make the single-pixel line in the middle of the Zero (0) character look more like a figure 8.  As I baseline everything to be presentable on real hardware on a real CRT from the era, not emulators that don;t emulate the screen behavior, the zero is edited to remove the pixels that mess up the appearance.