Fixed the glitch in the GTIA pixels
Stupid programming trick.   
When the code deletes the mothership from player memory the  index for the loop was incorrectly Y instead of X.   
At this point Y has the value 255, plus the forced offset from the base of Player 3 memory results in the code slapping a 0 into the LMS address for the GTIA pixels.  
Derp.
Putting X where it belongs fixed the stupid.