Added mountains and fixed other things
The mountains on the bottom of the screen are set up.  Some minor tweaks in various places.
DLIs are  running for the mountains, the static land, and the color change for the mode 2 status line at the bottom.
I will change the way the bumpers are drawn, and further redefine the character set to show rolling hills on the bottom ground line where the guns will be placed.