Tweaking the idle player
It takes several frames for the idle player to move up to the game play position.
There was a condition where the Player who presses the button late could have the gun motion still in progress when the Mothership animation starts which would stop the gun animation. 
So, the Idle Player input is separated from the Idle Player movement to the playing position to allow the animation part to continue being called during the Mothership animation.