WIP -- Starting work on Game Screen
There is a game screen executing and it isn't resulting in an immediate crash.
The VBI for this is calling the Flashing stars  loop (but nothing is yet appearing.... some bug, probably just lack of DLI running.)
The  VBI also calls the Player/Missile placements, and the ground scrolling code.   The ground code is clearly running, so things are looping where expected.
The land shifts between the Title screen and the Game screen, so there is an anomaly in the line counts between the two display lists.   They should be identical placements.