Resolved Altirra difference
Timing problems with DLI should be resolved in Altirra.
This was due to the last DLI for stars occasionally overlapping the next DLI.  
Removing one line of stars and replacing with a blank instruction solved that.

Status text was displaying on top of the idle guns on the title screen when tested on Windows Altirra, but do not appear in Linux atari800.   On real hardware the text does appear, so the problem is in atari800, and the solution is to REMOVE the statistics text from the display when not needed instead of just setting it black.   (Thanks to the author of Altirra for explaining this.)  The  correction of the text display  is still on the TODO list.

The code condition in this check-in should be the proper baseline for continuing the game logic.   The  Title and game screens should display  fine without DLI disaster on atari800, Altirra, and real hardware.  (The Count down timer still accelerated to make testing more convenient.)