Bang Bang Shoot Shoot
Lasers now function (but not collision detection).   
Some endurance testing in atari800 emulator with both guns against a bumper...  It  seems to function correctly regarding shooting and direction changes. 
Possible glitchiness remains.   Maybe it is an artifact of atari800 emulator.  Need to test in Altirra to figure out if it is real.  (When the game starts, sometimes it appears that a laser may flash for a frame  in the overscan area on the left side of the screen.   After a while the occasional laser image no longer appears.   strangeness.)