debugging the laser glitch in atari800
Cleaned up code to help figure out how HPOSP0/HPOSP1 is sometimes appearing in the left/right oversan area for one frame.   
Seems to be only two places that touch this -- the VBI sets laser HPOS, and the DLI sets HPOS for the guns.   
Aside from this, all DLI appear to function properly.
Tested timing of VBI, and Main line, and very little is happening -- Main actually finishes everything even before the top of the screen starts display, so this is not due to conflict between main and the DLI.
It may be specific to Atari800.   The behavior seems to go away after a minute or so... why....?   Need to check in Altirra.
Maybe something incomplete that gets filled in later as the game loops???  Doubtful.
Doubt