Partial working explosion.
Collision detection working.   Explosion appears at mothership position.
Did some violence to the laser code, so that the visual is messed up.   I know what I did wrong.  No time to fix it up at the moment.
The mothership position reset at the time of explosion is "working" .  The demo code to keep the mothership moving disagrees with the position change.   This will be fixed with the row, statistics, and scoring logic.