Almost working mothership
Found a missing transfer from New Y to Current Y for the mothership, which seemed to be what locked up the row to row progression.
Mothership mostly behaving much better.
Current issue is that on entry to the first line at the start of game the mothership is two lines high.  Though  I have currently hardcoded the correct starting positions for both in the game init section, something is interfering with the progression to the first line.