Mothership speed progression looks like it works..
Looks like the speed increases for the mothership are working.
"Looks Like", because now with the speedup  I haven't been able to play  it for testing up to 80 hits so far.   
I'll have to test with a different movement values, and/or debug to the screen to be sure this is indexing properly, and the right speeds are being chosen/added from the speed control table.