WIP -- working on row 22
Speed progression appears to be correct.  On to working on the last line (22) where the mothership pushes the ships off the screen....
Started minor tweaks supporting line 22.  Turn off the stats text line when the mothership reaches row 22, and conveniently, the $0 value for color can be used as a flag for exception logic rather than needing to compare the row to #22 all the time. 
Happened to notice the color for the mountains were not the right number of scan lines  per color:  3, 5, 3, 5, ... instead of 4, 4, 4,... DLI was short one WSYNC
Happened to also find that  the Mothership  Y coordinate location for the first row was set to 38 instead of 36.   ... I think this was accidental residue from debugging when the code was zipping the mothership along the vertical border of the screen instead of entering the line.