Erase Scores During Countdown
Playtesters commented that the giant mothership flies behind the scores/high score text.  
This is inevitable where Player/Missiles overlap ANTIC Mode 2 COLPF1 text pixels. No changes to priority (PRIOR) can fix this). 
Therefore, when the countdown starts the scores are removed from the screen to eliminate this inconsistent overlap (underlap(?))
The scores are redrawn when the countdown is over and the regular game starts.