New Font and other WIP
Updates are made to the font to allow for legible Mode 2 text. This includes rework of the characters displaying score.
Added DLI making gradient on the score line.   Similar gradient will (soon) appear for the status line.
WIP:
Added text line for the subtext  under the title graphics.
Added text  line for displaying OPTIONS and a line for descriptions  to the title screen.
Tweaked display list around to accommodate the extra lines.   This has resulted in some damage to the displays and the Game screen and Game over screens have dropped a scan line or two lower.   OOPS.   its on the short list of fixes.