Prettification
Player/Missile graphics are used to tint the scores to match the color of the active player(s).

An animated subtitle appears under the big title graphics providing some messages of emotional imapct. (ONE BUTTON -- ONE ALIEN -- ONE LIFE -- NO MERCY)

There is a new line added for Options, and a line for Options documentation.  Supporting this, the custom character set is re-done to include appropriate A-Z for ANTIC Mode 2 text.

(Scoring is temporarily broken.  Looks like this is due to the Player/Missile graphics overlays.  Looks like HITCLR needs to be done at a different time/place on screen.)