Fixed Player 2 Extra Credit
Every 10 shots (hmmmm) player 2 would be credited for shooting the mothership.
Turns out the code added to scale the mothership speed for PAL added another function layer which used the X register, and this resulted in changing the player in use by the code when the hit counter was decremented.  Consequently, it thought player 2 did the shot and so awarded the points.