Menu WIP again
I haz an epiphany.   Wasting lots of space repeating function addresses in an array sized for the menu options.   
Many of these values are directly related to the variable, not to the menu.
So, to eliminate several of these 45 element tables create a structure of configuration values that supply the configuration variable all the useful values specific to the config variable.  This is only 9 variables times the side of the config management values.   In exchange for this about 5 tables with addresses will be freed up.
So, reorganizing some things.   

In the middle of updates.  Do not Assemble this mess.