MENUTASTIC CHEAT MODE FEATURE
Cheat mode and supporting graphics are implemented.   
Cheat mode means the mothership will never reach the bottom line.
When cheat mode is on, there is no high score update, and the flashing stars graphic is replaced (alternately) with the images  for the characters C- H- E- A- T.
Some minor code optimizations included. (Direct image write for mothership in the game.)